UnBArqDsw/2020.1_G13_iGado#48 - Add blueprints to the project
