+++ docker-compose.yml
@@ -11,11 +11,10 @@ services:
     ports:
       - 5001:5000
     environment:
-      - FLASK_APP=project/__init__.py
       - FLASK_ENV=development
       - APP_SETTINGS=project.config.DevelopmentConfig
-      - DATABASE_URL=postgres://postgres:postgres@db:5432/example_dev
-      - DATABASE_TEST_URL=postgres://postgres:postgres@db:5432/example_test
+      - DATABASE_URL=postgres://postgres:postgres@db:5432/dev
+      - DATABASE_TEST_URL=postgres://postgres:postgres@db:5432/test
     depends_on: 
       - db
