+++ manage.py
@@ -2,10 +2,11 @@ from flask.cli import FlaskGroup
 import sys
 import unittest
 
-from project import app, db
+from project import create_app, db
+from project.api.models import Example
 
-
-cli = FlaskGroup(app)
+app = create_app()
+cli = FlaskGroup(create_app=create_app)
 
 
 @cli.command('recreate_db')
