+++ project/__init__.py
@@ -1,37 +1,30 @@
 import os
-from flask import Flask, jsonify
-from flask_restful import Resource, Api
+from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 
-# instantiate the app
-app = Flask(__name__)
-
-api = Api(app)
-
-# set config
-app_settings = os.getenv('APP_SETTINGS')
-app.config.from_object(app_settings)  
 
 # instantiate the db
-db = SQLAlchemy(app)
+db = SQLAlchemy()
 
-# model
-class Example(db.Model):  # new
-    __tablename__ = 'example'
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    username = db.Column(db.String(128), nullable=False)
-    email = db.Column(db.String(128), nullable=False)
-    active = db.Column(db.Boolean(), default=True, nullable=False)
+def create_app(script_info=None):
 
-    def __init__(self, username, email):
-        self.username = username
-        self.email = email
+    # instantiate the app
+    app = Flask(__name__)
 
-class Ping(Resource):
-    def get(self):
-        return {
-        'status': 'success',
-        'message': 'pong!'
-    }
+    # set config
+    app_settings = os.getenv('APP_SETTINGS')
+    app.config.from_object(app_settings)
 
-api.add_resource(Ping, '/ping')
+    # set up extensions
+    db.init_app(app)
+
+    # register blueprints
+    from project.api.example import example_blueprint
+    app.register_blueprint(example_blueprint)
+
+    # shell context for flask cli
+    @app.shell_context_processor
+    def ctx():
+        return {'app': app, 'db': db}
+
+    return app
