+++ project/db/create.sql
@@ -1,3 +1,3 @@
-CREATE DATABASE users_prod;
-CREATE DATABASE users_dev;
-CREATE DATABASE users_test;
+CREATE DATABASE prod;
+CREATE DATABASE dev;
+CREATE DATABASE test;
