+++ project/tests/base.py
@@ -1,6 +1,8 @@
 from flask_testing import TestCase
 
-from project import app, db
+from project import create_app, db
+
+app = create_app()
 
 
 class BaseTestCase(TestCase):
