+++ project/tests/test_config.py
@@ -4,7 +4,9 @@ import unittest
 from flask import current_app
 from flask_testing import TestCase
 
-from project import app
+from project import create_app
+
+app = create_app()
 
 
 class TestDevelopmentConfig(TestCase):
